Update vscode settings
- remove/replace some deprecated settings
- disable minimap
- enable text wrapping
- hide the terminal on startup

Note that hiding the terminal on startup only works in codespaces
AFTER the codespace has been created. i.e. if you start a codespace and
then reload the browser, or restart the codespace, the terminal will be hidden.
It won't be hidden immediately after a codespace starts, because of
setup that codespaces do, which we can't easily change.