Replace the terminal message
A codespace automatically starts up with a terminal open showing a
default "Welcome to codespaces" message. There's no offical way to
override that.
https://github.com/orgs/community/discussions/43534#discussioncomment-4652048

However, we can suppress the default message and replace the motd with
our own one, and make it appear with each new terminal. This means that
when a codespace starts up, finishes doing its startup things, and opens
a new terminal, users see a message about the tutorial, with a link to
the docs, and a note that they don't need and can close the terminal.

We also add a setting that makes links in the terminal clickable.